imap-test: client-state - balance appends and deletes If you have a high Dele=, then you'll delete messages way quicker than you can append them, and will never reach the target mailbox size.
Similarly, if Dele= is too low, you'll sit at the target mailbox size
and almost never actually append any new messages.

This patch balances the number of messages deleted against the rarity
of deletion (rarer => delete more each time), and the frequency of
appending (rarer => don't delete as many). In order to bounce around
the target, rather than always staying short, appending must be
permitted to over-fill the message box. The new append-blocking state
is 14% over the target (8/7ths), in tests, it never gets anywhere near
that high, no matter the parametrisation.

This will eventually hone in on the target size even in extreme corner
cases of Appe=1 Dele=1, Appe=1 Dele=100, and App=100 Dele=1, as well
as the default Appe=100 Dele=100.

More noise (and therefore a wider test of the functionality) is seen
when Dele= is very low.

Signed-off-by: Phil Carmody <phil@dovecot.fi>