user: Dangerously silence -Wformat-nonliteral
Taking a format string from the outside world normally suicide. And the
only thing worse than suicide is being persistently nagged about it.

However, the previous patch should mean that suicide is now impossible,
so this really is just nagging now. (Even if it's possible to open up
new holes in the future.)