conf: Sanitise format strings from tainted sources
Make sure the '%' expressions only have int type, and the quantity is no
greater than the number of ints we'll actually be passing. %i is allowed
in case people forget it should be %d. Only /%[0-9]*[id]/ is considered
valid for formatting ints at this point.

This makes the previous dangerous patch that touched the username parsing
not so dangerous any more.