Actually, never continue creating connection to a same user.
Even with 10% chance of doing it, it's a lot of them with high enough
client counts. It also makes much more sense to just do it randomly. If
you want to test multiple connections to the same user, just user a
low enough user count and high enough client count.