tests: If test is skipped, LOGOUT all connections cleanly
This allows cleanly shutting down the IMAP server immediately after imaptest
is run. Otherwise there could be some errors logged when some connection is
still in the middle of logging in. This was happening especially because
urlauth tests are the last tests to be run and they use multiple connections.